datos dummy para los servicios rest

+++ database/migrations/2019_11_07_19_create_cost_table.php
@@ -14,18 +14,18 @@ class CreateCostTable extends Migration
     public function up()
     {
         Schema::create('cost', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->increments('id');            
             $table->unsignedInteger('costad_id');
             $table->foreign('costad_id')->references('id')->on('cost_ad');            
             $table->unsignedInteger('project_id');
             $table->foreign('project_id')->references('id')->on('projects');            
             $table->unsignedInteger('costdate_id');
-            $table->foreign('costdate_id')->references('id')->on('cost_date');                        
+            $table->foreign('costdate_id')->references('id')->on('cost_date');             
             $table->integer('Horas_Code')->nullable();
             $table->integer('Horas_Script')->nullable();
             $table->integer('Horas_DP')->nullable();
             $table->double('ValorHora')->nullable();
-            $table->double('CostoTotal')->nullable();
+            $table->double('CostoTotal')->nullable();            
             $table->timestamps();
         });
     }
